// src/app/data/models/tag.ts
export enum Tag {
  Rolling = 'rolling',
  Stable = 'stable',
  Immutable = 'immutable',
  Beginner = 'beginner',
  Terminal = 'terminal',
  Diy = 'diy',
  Ricing = 'ricing',
  Polished = 'polished',
  Gaming = 'gaming',
  Lightweight = 'lightweight',
  Privacy = 'privacy',
  Server = 'server',

  SystemdFree = 'systemd-free',
  Libre = 'libre',
  Corporate = 'corporate',
  WillBreak = 'will-break',
  SecurityTools = 'security-tools',
  LiveUsb = 'live-usb',
  Arm = 'arm',
  WindowsLike = 'windows-like',
  MacLike = 'mac-like',
  Obscure = 'obscure',
}